requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
greenlet==3.1.0
idna==3.8
psycopg2-binary==2.9.9
requests==2.32.3
SQLAlchemy==2.0.34
typing_extensions==4.12.2
urllib3==2.2.2
Flask_Login==0.6.3
Werkzeug==3.0.0
Flask==2.3.2
flask_sqlalchemy==3.1.1
python_dotenv==1.0.1
Flask-Limiter==3.0.0
Bootstrap_Flask==2.3.3
gunicorn==21.2.0
pymysql
